requirements.txt
python-binance
websocket-client
aiofiles
pandas
asyncio
aiocsv
numpy
mplfinance
tweepy
matplotlib
